requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
whitenoise==6.1.0

djangorestframework~=3.13.1